--- plugins/variants/src/LinearVariantDisplay/configSchema.test.js
@@ -1,55 +1,0 @@
-import BoxRendererType from '@jbrowse/core/pluggableElementTypes/renderers/BoxRendererType'
-import Plugin from '@jbrowse/core/Plugin'
-import PluginManager from '@jbrowse/core/PluginManager'
-import PileupRenderer from '@jbrowse/plugin-alignments/src/PileupRenderer'
-import {
-  configSchema as svgFeatureRendererConfigSchema,
-  ReactComponent as SvgFeatureRendererReactComponent,
-} from '@jbrowse/plugin-svg/src/SvgFeatureRenderer'
-import configSchemaFactory from './configSchema'
-
-// mock warnings to avoid unnecessary outputs
-beforeEach(() => {
-  jest.spyOn(console, 'warn').mockImplementation(() => {})
-})
-
-afterEach(() => {
-  console.warn.mockRestore()
-})
-
-class PileupRendererPlugin extends Plugin {
-  install(pluginManager) {
-    PileupRenderer(pluginManager)
-  }
-}
-
-class SvgFeatureRendererPlugin extends Plugin {
-  install(pluginManager) {
-    pluginManager.addRendererType(
-      () =>
-        new BoxRendererType({
-          name: 'SvgFeatureRenderer',
-          ReactComponent: SvgFeatureRendererReactComponent,
-          configSchema: svgFeatureRendererConfigSchema,
-          pluginManager,
-        }),
-    )
-  }
-}
-
-test('has a viewType attr', () => {
-  const configSchema = configSchemaFactory(
-    new PluginManager([
-      new PileupRendererPlugin(),
-      new SvgFeatureRendererPlugin(),
-    ])
-      .createPluggableElements()
-      .configure(),
-  )
-  const config = configSchema.create({
-    type: 'LinearVariantDisplay',
-    displayId: 'displayId0',
-    name: 'Zonker Display',
-  })
-  expect(config.type).toEqual('LinearVariantDisplay')
-})
