+++ plugins/wiggle/src/DensityRenderer/DensityRenderer.test.ts
@@ -6,10 +6,13 @@ import ReactComponent from '../WiggleRendering'
 
 import { Image, createCanvas } from 'canvas'
 
+// @ts-ignore
 global.nodeImage = Image
+
+// @ts-ignore
 global.nodeCreateCanvas = createCanvas
 
-const pluginManager = {}
+const pluginManager = {} as any
 const renderer = new DensityRenderer({
   name: 'DensityRenderer',
   ReactComponent,
@@ -48,6 +51,7 @@ test('inverted mode and reversed', async () => {
   }
 
   const res = await renderToAbstractCanvas(1000, 200, renderProps, ctx =>
+    // @ts-ignore
     renderer.draw(ctx, renderProps),
   )
   expect(res).toMatchSnapshot({
